Skip NDK download if it's already present in cache

+++ android/buildAndroidBOINC-CI.sh
@@ -129,8 +129,12 @@ fi
 export CURL_SRC=$BUILD_DIR/curl-${CURL_VERSION}
 
 export NDK_VERSION=17c
-wget -O /tmp/ndk.zip https://dl.google.com/android/repository/android-ndk-r${NDK_VERSION}-linux-x86_64.zip
-unzip -qq /tmp/ndk.zip -d $HOME
+export NDK_FLAGFILE="$PREFIX/${arch}/NDK-${NDK_VERSION}_done"
+if [ ! -e "${NDK_FLAGFILE}" ]; then
+    rm -rf "${PREFIX}/${arch}"
+    wget -O /tmp/ndk.zip https://dl.google.com/android/repository/android-ndk-r${NDK_VERSION}-linux-x86_64.zip
+    unzip -qq /tmp/ndk.zip -d $HOME
+fi
 export NDK_ROOT=$HOME/android-ndk-r${NDK_VERSION}
 
 export ANDROID_TC=$PREFIX
@@ -138,6 +142,7 @@ export ANDROID_TC=$PREFIX
 case "$arch" in
     "arm")
         ./build_androidtc_arm.sh
+        touch "${NDK_FLAGFILE}"
         ./build_openssl_arm.sh
         ./build_curl_arm.sh
         ./build_boinc_arm.sh
@@ -145,6 +150,7 @@ case "$arch" in
     ;;
     "arm64")
         ./build_androidtc_arm64.sh
+        touch "${NDK_FLAGFILE}"
         ./build_openssl_arm64.sh
         ./build_curl_arm64.sh
         ./build_boinc_arm64.sh
@@ -152,6 +158,7 @@ case "$arch" in
     ;;
     "x86")
         ./build_androidtc_x86.sh
+        touch "${NDK_FLAGFILE}"
         ./build_openssl_x86.sh
         ./build_curl_x86.sh
         ./build_boinc_x86.sh
@@ -159,6 +166,7 @@ case "$arch" in
     ;;
     "x86_64")
         ./build_androidtc_x86_64.sh
+        touch "${NDK_FLAGFILE}"
         ./build_openssl_x86_64.sh
         ./build_curl_x86_64.sh
         ./build_boinc_x86_64.sh
